+++ temba/channels/types/viber_public/tests.py
@@ -81,7 +81,7 @@ class ViberPublicTypeTest(TembaTest, CRUDLTestMixin):
             update_url,
             allow_viewers=False,
             allow_editors=True,
-            form_fields={"name": "Viber", "alert_email": None, "welcome_message": ""},
+            form_fields={"name": "Viber", "welcome_message": ""},
         )
 
         self.assertUpdateSubmit(
@@ -96,11 +96,7 @@ class ViberPublicTypeTest(TembaTest, CRUDLTestMixin):
             update_url,
             allow_viewers=False,
             allow_editors=True,
-            form_fields={
-                "name": "Updated",
-                "alert_email": None,
-                "welcome_message": "Welcome, please subscribe for more",
-            },
+            form_fields={"name": "Updated", "welcome_message": "Welcome, please subscribe for more"},
         )
 
         # read page has link to update page
@@ -110,7 +106,7 @@ class ViberPublicTypeTest(TembaTest, CRUDLTestMixin):
         self.login(self.customer_support, choose_org=self.org)
         response = self.client.get(update_url)
         self.assertEqual(
-            ["name", "alert_email", "log_policy", "welcome_message", "loc"],
+            ["name", "log_policy", "welcome_message", "loc"],
             list(response.context["form"].fields.keys()),
         )