+++ temba/channels/types/whatsapp/views.py
@@ -389,9 +389,7 @@ class Connect(ChannelTypeMixin, OrgPermsMixin, SmartFormView):
     title = "Connect WhatsApp"
 
     def has_org_perm(self, permission):
-        if self.org:
-            return self.get_user().is_beta  # only beta users are allowed
-        return False  # pragma: no cover
+        return self.get_user().is_beta  # only beta users are allowed
 
     def pre_process(self, request, *args, **kwargs):
         session_token = self.request.session.get(self.channel_type.SESSION_USER_TOKEN, None)