+++ temba/mailroom/tests.py
@@ -897,12 +897,6 @@ class EventTest(TembaTest):
                     "text": "Hello",
                     "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
                     "quick_replies": ["yes", "no"],
-                    "created_by": {
-                        "id": self.agent.id,
-                        "email": "agent@nyaruka.com",
-                        "first_name": "Agnes",
-                        "last_name": "",
-                    },
                 },
                 "created_by": {
                     "id": self.agent.id,
@@ -910,6 +904,7 @@ class EventTest(TembaTest):
                     "first_name": "Agnes",
                     "last_name": "",
                 },
+                "optin": None,
                 "status": "E",
                 "logs_url": f"/channels/{str(self.channel.uuid)}/logs/msg/{msg_out.id}/",
             },
@@ -930,9 +925,9 @@ class EventTest(TembaTest):
                     "urn": "tel:+250979111111",
                     "text": "Hello",
                     "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
-                    "created_by": None,
                 },
                 "created_by": None,
+                "optin": None,
                 "status": "F",
                 "failed_reason": "D",
                 "failed_reason_display": "No suitable channel found",
@@ -953,7 +948,6 @@ class EventTest(TembaTest):
                     "urn": "tel:+250979111111",
                     "text": "Hello",
                     "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
-                    "created_by": None,
                 },
                 "created_by": None,
                 "status": "S",
@@ -984,11 +978,105 @@ class EventTest(TembaTest):
                     "first_name": "Andy",
                     "last_name": "",
                 },
+                "optin": None,
                 "status": "S",
                 "recipient_count": 2,
                 "logs_url": f"/channels/{str(self.channel.uuid)}/logs/msg/{msg_out2.id}/",
             },
             Event.from_msg(self.org, self.admin, msg_out2),
+        )
+
+        # create a broadcast that was sent with an opt-in
+        optin = self.create_optin("Polls")
+        bcast2 = self.create_broadcast(self.admin, "Hi there", contacts=[contact1, contact2], optin=optin)
+        msg_out3 = bcast2.msgs.filter(contact=contact1).get()
+
+        self.assertEqual(
+            {
+                "type": "broadcast_created",
+                "created_on": matchers.ISODate(),
+                "translations": {"und": {"text": "Hi there"}},
+                "base_language": "und",
+                "msg": {
+                    "uuid": str(msg_out3.uuid),
+                    "id": msg_out3.id,
+                    "urn": "tel:+250979111111",
+                    "text": "Hi there",
+                    "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
+                },
+                "created_by": {
+                    "id": self.admin.id,
+                    "email": "admin@nyaruka.com",
+                    "first_name": "Andy",
+                    "last_name": "",
+                },
+                "optin": {"uuid": str(optin.uuid), "name": "Polls"},
+                "status": "S",
+                "recipient_count": 2,
+                "logs_url": f"/channels/{str(self.channel.uuid)}/logs/msg/{msg_out3.id}/",
+            },
+            Event.from_msg(self.org, self.admin, msg_out3),
+        )
+
+        # create a message that was an opt-in request
+        msg_out4 = self.create_optin_request(contact1, self.channel, optin)
+        self.assertEqual(
+            {
+                "type": "optin_requested",
+                "created_on": matchers.ISODate(),
+                "optin": {"uuid": str(optin.uuid), "name": "Polls"},
+                "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
+                "urn": "tel:+250979111111",
+                "created_by": None,
+                "status": "S",
+                "logs_url": f"/channels/{str(self.channel.uuid)}/logs/msg/{msg_out4.id}/",
+            },
+            Event.from_msg(self.org, self.admin, msg_out4),
+        )
+
+    def test_from_channel_event(self):
+        self.create_contact("Jim", phone="+250979111111")
+
+        event1 = self.create_channel_event(
+            self.channel, "tel:+250979111111", ChannelEvent.TYPE_CALL_IN, extra={"duration": 5}
+        )
+
+        self.assertEqual(
+            {
+                "type": "channel_event",
+                "created_on": matchers.ISODate(),
+                "event": {
+                    "type": "mo_call",
+                    "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
+                    "duration": 5,
+                },
+                "channel_event_type": "mo_call",  # deprecated
+                "duration": 5,  # deprecated
+            },
+            Event.from_channel_event(self.org, self.admin, event1),
+        )
+
+        optin = self.create_optin("Polls")
+        event2 = self.create_channel_event(
+            self.channel,
+            "tel:+250979111111",
+            ChannelEvent.TYPE_OPTIN,
+            extra={"optin_id": optin.id, "optin_name": "Polls"},
+        )
+
+        self.assertEqual(
+            {
+                "type": "channel_event",
+                "created_on": matchers.ISODate(),
+                "event": {
+                    "type": "optin",
+                    "channel": {"uuid": str(self.channel.uuid), "name": "Test Channel"},
+                    "optin": {"uuid": str(optin.uuid), "name": "Polls"},
+                },
+                "channel_event_type": "optin",  # deprecated
+                "duration": None,  # deprecated
+            },
+            Event.from_channel_event(self.org, self.admin, event2),
         )
 
     def test_from_flow_run(self):