+++ temba/notifications/incidents/builtin.py
@@ -1,6 +1,31 @@
+from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
 from ..models import Incident, IncidentType
+
+
+class ChannelDisconnectedIncidentType(IncidentType):
+    """
+    Android channel appears to be disconnected.
+    """
+
+    slug = "channel:disconnected"
+    title = _("Channel Disconnected")
+
+    @classmethod
+    def get_or_create(cls, channel):
+        """
+        Creates a channel disconnected incident if one is not already ongoing
+        """
+        return Incident.get_or_create(
+            channel.org, ChannelDisconnectedIncidentType.slug, scope=str(channel.id), channel=channel
+        )
+
+    def get_notification_scope(self, incident) -> str:
+        return str(incident.channel.id)
+
+    def get_notification_target_url(self, incident) -> str:
+        return reverse("channels.channel_read", args=[str(incident.channel.uuid)])
 
 
 class OrgFlaggedIncidentType(IncidentType):