+++ temba/triggers/views.py
@@ -293,7 +293,7 @@ class TriggerCRUDL(SmartCRUDL):
             def add_section(name, url, icon):
                 formax.add_section(name, reverse(url), icon=icon, action="redirect", button=_("New Trigger"))
 
-            org_schemes = self.org.get_schemes(Channel.ROLE_RECEIVE)
+            org_schemes = self.request.org.get_schemes(Channel.ROLE_RECEIVE)
 
             add_section("trigger-keyword", "triggers.trigger_create_keyword", "flow")
             add_section("trigger-register", "triggers.trigger_create_register", "group")
@@ -301,7 +301,7 @@ class TriggerCRUDL(SmartCRUDL):
             add_section("trigger-schedule", "triggers.trigger_create_schedule", "calendar")
             add_section("trigger-inboundcall", "triggers.trigger_create_inbound_call", "incoming_call")
 
-            if self.org.channels.filter(is_active=True, channel_type=AndroidType.code).exists():
+            if self.request.org.channels.filter(is_active=True, channel_type=AndroidType.code).exists():
                 add_section("trigger-missedcall", "triggers.trigger_create_missed_call", "missed_call")
 
             if ContactURN.SCHEMES_SUPPORTING_NEW_CONVERSATION.intersection(org_schemes):
@@ -313,8 +313,8 @@ class TriggerCRUDL(SmartCRUDL):
             add_section("trigger-closed-ticket", "triggers.trigger_create_closed_ticket", "agent")
 
             if self.request.user.is_staff:
-                add_section("trigger-opt-in", "triggers.trigger_create_opt_in", "restore")
-                add_section("trigger-opt-out", "triggers.trigger_create_opt_out", "error")
+                add_section("trigger-opt-in", "triggers.trigger_create_opt_in", "optin")
+                add_section("trigger-opt-out", "triggers.trigger_create_opt_out", "optout")
 
     class BaseCreate(OrgPermsMixin, ComponentFormMixin, SmartCreateView):
         trigger_type = None